Add gridCreate()
Added the function instructions for gridCreate(). This function is called when the event listener fires on a change from gridSelect. The grid creates a new div in a loop and adds it as a child with the class .box in the gridContainer. This is still a WIP with the core function being present.