Add gridErase() function
Added a new function which checks for children being present in gridContainer, it then proceeds to remove all children. This function is called when a change is made and createGrid is called. This way the board is clean and only the selected amount of grids are ever deployed to the gridContainer.