Add changeColor(), and event listener
Added a new function which generates a random hex value color. This color is then applied via an event listener. The event listener and nodelist are update on a 500ms basis so that each time the new divs are deployed they are detected and listened on. The box.style.background color are updated via string template which references the changeColor function