Add styling to html structure
Added the styling for the calculator. This is a final submission for the styling unless changes are needed later. The theme is a dark/green with hover and active affects applied to the buttons. Eventually an effort to make the calculator responsive may be attempted. As of now it is fixed for a large display.